refactor: lazy load footer nav bar
